requirements.txt
# UOB Risk & Regulatory IT Survey Application
# Python 3.11+ required

# Core Framework
streamlit>=1.28.0
streamlit-option-menu>=0.3.2

# AI/LLM
openai>=1.0.0

# Database
pymongo>=4.5.0
motor>=3.3.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# PDF Generation
fpdf2>=2.7.0
reportlab>=4.0.0
weasyprint>=59.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Development (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
streamlit==1.28.1
configparser==6.0.0
pymongo==4.6.0
openai==1.3.9
python-dotenv==1.0.0
fpdf2==2.7.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
markdown==3.5.1
plotly==5.18.0
